Updated current repositories and changed averageLifespan datatype

+++ autotask/src/main/java/z21/autotask/orm/Species.java
@@ -33,7 +33,7 @@ public class Species {
     private String foodType;
 
     @Column(name = "average_lifespan") 
-    private Float averageLifespan;
+    private Integer averageLifespan;
 
     @OneToMany(mappedBy = "speciesId")
     private List<Animal> animals;
